Minor format and typo fixes in README.md
Not a detailed look and it was a look with bleary eyes but a few things
did 'pop out' and they should now be fixed. There might (or might not)
be other things. One issue was a formatting error due to old style
writing (that I'm also guilty of doing - ` followed by ') but which
doesn't work very well in markdown.